dumper: Randomize Branch Names
(1). This is needed in order to dump custom-roms, as multiple builds of the same device might use the same fingerprint, thus not pushing the firmware, or git notifying to use "-f" flag, which is not what we want.
(2). We aren't hosting our dumps on a custom server, so storage isn't much of a headache as of now.